+++ shell.nix
@@ -1,4 +1,4 @@
 { pkgs ? import <nixpkgs> {} }:
 pkgs.mkShell {
-  nativeBuildInputs = with pkgs; [ picocom grabserial clang wxmaxima gcc ];
+  nativeBuildInputs = with pkgs; [ picocom grabserial clang wxmaxima gcc libelf ];
 }